refactor: move parsing execution args outside file manager class
